# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/users": {
      "post": {
        "description": "Create a new user in the system. Only administrators can create users.",
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized access"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "409": {
            "description": "Email already exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new user",
        "tags": [
          "👥 Usuários"
        ]
      },
      "get": {
        "description": "Retrieve a paginated list of all users. Accessible by administrators and moderators.",
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page (default: 10, max: 100)",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term for name or email",
            "schema": {
              "example": "john",
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "description": "Filter by user role",
            "schema": {
              "enum": [
                "user",
                "admin",
                "super-admin",
                "guest"
              ],
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filter by user active status",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access"
          },
          "403": {
            "description": "Insufficient permissions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all users",
        "tags": [
          "👥 Usuários"
        ]
      }
    },
    "/v1/users/stats": {
      "get": {
        "description": "Get comprehensive statistics about users in the system. Only accessible by administrators.",
        "operationId": "UsersController_getStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserStatsResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access"
          },
          "403": {
            "description": "Insufficient permissions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user statistics",
        "tags": [
          "👥 Usuários"
        ]
      }
    },
...
    "/v1/users/{id}": {
      "get": {
        "description": "Retrieve a specific user by their ID. Accessible by administrators and moderators.",
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user by ID",
        "tags": [
          "👥 Usuários"
        ]
      },
      "patch": {
        "description": "Update a specific user by their ID. Only accessible by administrators.",
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized access"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "User not found"
          },
          "409": {
            "description": "Email already exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update user by ID",
        "tags": [
          "👥 Usuários"
        ]
      },
      "delete": {
        "description": "Permanently delete a user from the system. Only accessible by administrators. This action cannot be undone.",
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted successfully"
          },
          "401": {
            "description": "Unauthorized access"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "User not found"
          },
          "409": {
            "description": "Cannot delete user with active dependencies"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete user by ID",
        "tags": [
          "👥 Usuários"
        ]
      }
    },
...
    "/v1/plans/admin/stats": {
      "get": {
        "description": "Retorna estatísticas gerais dos planos cadastrados. Requer permissões de administrador.",
        "operationId": "PlansController_getStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Estatísticas retornadas com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "number",
                      "example": 10
                    },
                    "active": {
                      "type": "number",
                      "example": 8
                    },
                    "inactive": {
                      "type": "number",
                      "example": 2
                    },
                    "featured": {
                      "type": "number",
                      "example": 3
                    },
                    "byType": {
                      "type": "object",
                      "example": {
                        "free": 2,
                        "premium": 5,
                        "enterprise": 3
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Acesso negado - permissões insuficientes"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter estatísticas dos planos",
        "tags": [
          "📋 Planos"
        ]
      }
    },
...
    "/v1/admin/stats": {
      "get": {
        "description": "Retorna estatísticas detalhadas do sistema incluindo usuários, receita, créditos e performance",
        "operationId": "AdminController_getSystemStats",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Data de início para as estatísticas",
            "schema": {
              "example": "2025-01-01",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "Data de fim para as estatísticas",
            "schema": {
              "example": "2025-12-31",
              "type": "string"
            }
          },
          {
            "name": "period",
            "required": false,
            "in": "query",
            "description": "Período das estatísticas",
            "schema": {
              "example": "monthly",
              "type": "string",
              "enum": [
                "daily",
                "weekly",
                "monthly",
                "yearly"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Estatísticas obtidas com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminStatsResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token não fornecido ou inválido"
          },
          "403": {
            "description": "Usuário não tem permissões administrativas"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter estatísticas do sistema",
        "tags": [
          "⚙️ Administração"
        ]
      }
    },
    "/v1/admin/users": {
      "get": {
        "description": "Lista usuários do sistema com opções de filtro e paginação",
        "operationId": "AdminController_getUsers",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filtrar por status",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "inactive",
                "pending",
                "suspended"
              ]
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "description": "Filtrar por role",
            "schema": {
              "type": "string",
              "enum": [
                "user",
                "admin",
                "moderator",
                "super_admin"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Buscar por nome ou email",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Página da listagem",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Itens por página",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuários obtida com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar usuários",
        "tags": [
          "⚙️ Administração"
        ]
      }
    },
    "/v1/admin/users/{id}": {
      "get": {
        "description": "Retorna informações detalhadas de um usuário específico",
        "operationId": "AdminController_getUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do usuário",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter usuário por ID",
        "tags": [
          "⚙️ Administração"
        ]
      },
      "put": {
        "description": "Atualiza informações de um usuário específico",
        "operationId": "AdminController_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do usuário",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos fornecidos"
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar usuário",
        "tags": [
          "⚙️ Administração"
        ]
      }
    },
    "/v1/admin/users/bulk-action": {
      "post": {
        "description": "Executa uma ação em múltiplos usuários simultaneamente",
        "operationId": "AdminController_bulkAction",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminBulkActionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ação executada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "affected": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ação inválida ou dados incorretos"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ação em lote",
        "tags": [
          "⚙️ Administração"
        ]
      }
    },
    "/v1/admin/dashboard": {
      "get": {
        "description": "Retorna dados resumidos para o dashboard administrativo",
        "operationId": "AdminController_getDashboard",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Dados do dashboard obtidos com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "overview": {
                      "type": "object",
                      "properties": {
                        "totalUsers": {
                          "type": "number"
                        },
                        "activeUsers": {
                          "type": "number"
                        },
                        "totalRevenue": {
                          "type": "number"
                        },
                        "totalCreditsConsumed": {
                          "type": "number"
                        }
                      }
                    },
                    "recentActivity": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "timestamp": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "userId": {
                            "type": "string"
                          },
                          "userName": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "systemHealth": {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string"
                        },
                        "uptime": {
                          "type": "number"
                        },
                        "memoryUsage": {
                          "type": "number"
                        },
                        "responseTime": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Dashboard administrativo",
        "tags": [
          "⚙️ Administração"
        ]
      }
    },
    "/v1/admin/export/users": {
      "get": {
        "description": "Exporta dados de usuários em formato CSV",
        "operationId": "AdminController_exportUsers",
        "parameters": [
          {
            "name": "format",
            "required": false,
            "in": "query",
            "description": "Formato do export (csv, xlsx)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Export iniciado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "downloadUrl": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Exportar usuários",
        "tags": [
          "⚙️ Administração"
        ]
      }
    },
...
  },
  "info": {
    "title": "🧠 NeuralContent API",
    "description": "\n## API completa para plataforma de geração de conteúdo com IA\n\n### 🔐 Autenticação\nEsta API utiliza **JWT Bearer tokens**. Para autenticar suas requisições:\n```\nAuthorization: Bearer {seu_jwt_token}\n```\n\n### 📊 Rate Limiting\n- **Usuários Free:** 100 req/min\n- **Usuários Pro:** 500 req/min  \n- **Usuários Enterprise:** 1000 req/min\n- **Administradores:** Sem limite\n\n### 🔄 Versionamento\n- **Versão Atual:** v1.0.0\n- **URL Base:** `/v1`\n- **Suporte:** Versionamento semântico (SemVer)\n\n### 📈 Funcionalidades Principais\n- **Autenticação e Autorização** (JWT + Roles)\n- **Gestão de Usuários** (CRUD + Perfis)\n- **Sistema de Planos** (Free, Pro, Enterprise)\n- **Gestão de Créditos** (Compra, Consumo, Histórico)\n- **Processamento de Pagamentos** (Stripe Integration)\n- **Monitoramento e Logs** (Health Checks + Analytics)\n\n### 🛡️ Segurança\n- Criptografia de senhas (bcrypt)\n- Validação de entrada (class-validator)\n- Proteção CORS configurável\n- Headers de segurança (Helmet)\n- Rate limiting por usuário\n    ",
    "version": "1.0.0",
    "contact": {
      "name": "NeuralContent Support",
      "url": "https://neuralcontent.com/support",
      "email": "api-support@neuralcontent.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "termsOfService": "https://neuralcontent.com/terms"
  },
...
  "servers": [
    {
...
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "JWT token obtido via /auth/login. Formato: Bearer {token}",
        "in": "header"
      },
      "ApiKey-auth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key para integração de sistemas externos"
      }
    },
    "schemas": {
...
      "UserListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "description": "Lista de usuários",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserResponseDto"
            }
          },
          "meta": {
            "type": "object",
            "description": "Metadados de paginação",
            "example": {
              "total": 150,
              "page": 1,
              "limit": 10,
              "totalPages": 15
            }
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
    }
  }
}